airport_backend/requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.3
packaging==24.1
propcache==0.2.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pyTelegramBotAPI==4.23.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
telebot==0.0.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
djangorestframework-simplejwt==5.3.1
